hackathonfinal-main/mcp_servers/live_events_server/pyproject.toml
[project]
name = "mcp-live-events-server"
version = "0.1.0"
description = "MCP server for Ticketmaster live events data"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []